Added test protoplugin utils
